Extract function encode_attributes to src/libindy/credentials
