+++ aries_vcx/src/libindy/mod.rs
@@ -2,3 +2,4 @@ pub mod proofs;
 pub mod utils;
 pub mod credential_def;
 pub mod schema;
+pub mod credentials;
