+++ aries_vcx/src/libindy/utils/anoncreds.rs
@@ -667,7 +667,7 @@ pub mod test_utils {
 
     use crate::{libindy, settings};
     use crate::libindy::credential_def::{CredentialDef, CredentialDefConfigBuilder, RevocationDetailsBuilder};
-    use crate::handlers::issuance::issuer::utils::encode_attributes;
+    use crate::libindy::credentials::encode_attributes;
     use crate::utils::constants::{TAILS_DIR, TEST_TAILS_URL};
     use crate::utils::get_temp_dir_path;
